Clean up code and move SparkOnEks to non plugin
