+++ cmd/punch/uninstall.go
@@ -23,6 +23,7 @@ import (
 	_ "github.com/datapunchorg/punch/pkg/topologyimpl/eks"
 	_ "github.com/datapunchorg/punch/pkg/topologyimpl/kafkabridge"
 	_ "github.com/datapunchorg/punch/pkg/topologyimpl/kafkaonmsk"
+	_ "github.com/datapunchorg/punch/pkg/topologyimpl/sparkoneks"
 	"github.com/spf13/cobra"
 )
