+++ pkg/topologyimpl/hivemetastore/topologyhandler.go
@@ -19,7 +19,6 @@ package main
 import (
 	"fmt"
 	"log"
-	"regexp"
 	"strings"
 
 	"github.com/datapunchorg/punch/pkg/awslib"
@@ -30,14 +29,7 @@ import (
 	"gopkg.in/yaml.v3"
 )
 
-var nonAlphanumericRegexp *regexp.Regexp
-
 func init() {
-	var err error
-	nonAlphanumericRegexp, err = regexp.Compile("[^a-zA-Z]+")
-	if err != nil {
-		panic(err)
-	}
 }
 
 type TopologyHandler struct {
