+++ pkg/topologyimpl/kafkabridge/topologyhandler.go
@@ -25,20 +25,11 @@ import (
 	"github.com/datapunchorg/punch/pkg/topologyimpl/kafkaonmsk"
 	"gopkg.in/yaml.v3"
 	"log"
-	"regexp"
 	"time"
 )
 
-var nonAlphanumericRegexp *regexp.Regexp
-
 func init() {
 	framework.DefaultTopologyHandlerManager.AddHandler(KindKafkaTopology, &TopologyHandler{})
-
-	var err error
-	nonAlphanumericRegexp, err = regexp.Compile("[^a-zA-Z]+")
-	if err != nil {
-		panic(err)
-	}
 }
 
 type createTopicRequest struct {
